Directories can now save their entries to multiple blocks
Before: when a new dir_entry was added to a directory, it was added to its first block only. There was no way of telling when that block was full of entries (a block has 4096 bytes) and needed a new block.

Now: if a dir_entry doesn't fit in a block (because it is full of entries), then the system will assign a new block to the directory, update the directory's inode (to add the new pointer), and save the dir_entry in the new block.

This should have the same functionality as before (I think)